Xu's Lab 1 update
Xu's Lab 1 update

+++ Lab1/Sources/FIFO.h
@@ -17,8 +17,8 @@
 // FIFO structure
 typedef struct
 {
-  UINT16 Start, End;
-  UINT16 volatile NbBytes;
+  UINT8 Start, End;
+  UINT8 volatile NbBytes;
   UINT8 Buffer[FIFO_SIZE];
 } TFIFO;
